# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/sets": {
            "get": {
                "description": "Returns a list of all LEGO sets in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lego"
                ],
                "summary": "Get all LEGO sets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sets/{id}": {
            "get": {
                "description": "Retrieves a LEGO set by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lego"
                ],
                "summary": "Get a LEGO set by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LEGO Set ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}